updated purge.js so you can delete a single message from designated user
