+++ ChaosBot.js
@@ -4,6 +4,7 @@ const client = new Discord.Client();
 const config = require("./config.json");
 var helpFileLoc = 'helpfile.txt'
 var poll = require("./poll.js");
+var purge = require("./purge.js");
 
 
 client.on("ready", () => {
@@ -80,51 +81,7 @@ const command = args.shift().toLowerCase();
 
 // Start Purge
     if (command === "purge") {
-
-      if(message.author.id === config.ownerID || message.author.id === config.James || message.author.id === config.Jesse) {
-
-        var excessDeletion = false;
-        const user = message.mentions.users.first();
-        let amount = parseInt(args[args.length - 1],10);
-
-        // Invalid input errors
-        if (isNaN(amount) || amount < 2) return message.reply('Must specify an amount (2-50) to delete!');
-
-        if (user && user.id === message.author.id) {amount = amount + 1;} else if (!user) {amount = amount + 1;}
-        console.log(` Preparing to delete ${amount} messages (including sent message)`);
-
-        // Prevent excess deletion
-        if (amount > 51) {excessDeletion = true;}
-        if (excessDeletion === true) {
-          amount = 51;
-          console.log(" Reduced to 51 (actually 50)");
-        }
-
-        // If User Specified
-        if (user) {
-            message.channel.fetchMessages({
-              limit: 100,
-            }).then((messages) => {
-               const filterBy = user ? user.id : Client.user.id;
-               messages = messages.filter(m => m.author.id === filterBy).array().slice(0, amount);
-               if (messages.length >= 2) {
-                 message.channel.bulkDelete(messages).catch(error => console.log(error.stack));
-                } else return message.reply(' not enough messages from specified user')
-            });
-          } else {
-        // If No User Specified
-            message.channel.fetchMessages({
-              limit: amount,
-            }).then((messages) => {
-               message.channel.bulkDelete(messages).catch(error => console.log(error.stack));
-            });
-          }
-        // Warning About Being Too Keen
-          if (excessDeletion === true) {
-            message.channel.send('Too big, reduced to 50.');
-          }
-          excessDeletion = false;
-      } else return message.reply('Not Authorized To Purge!');
+      purge.run(client, message, config, args);
     }
 // End Purge
